Seperated generated attributes into samller, seperate files
