+++ src/componants/Description.svelte
@@ -12,8 +12,29 @@
 {:else}
 	<Paper color="primary" variant="outlined" class="mdc-theme--primary">
 		<div class="characterDescription">
-			{#if character && character.siblings && character.caretakerStatus}
+			{#if character}
 				<LayoutGrid>
+					<Cell class="section">
+						<h2>Basic Info</h2>
+						<div class="subsection">
+							<Content>
+								<b>Name:</b>
+								{character.name}
+							</Content>
+							<Content>
+								<b>Age:</b>
+								{character.age}
+							</Content>
+							<Content>
+								<b>Gender:</b>
+								{character.gender}
+							</Content>
+							<Content>
+								<b>Race:</b>
+								{character.race}
+							</Content>
+						</div>
+					</Cell>
 					<Cell class="section">
 						<h2>Origin</h2>
 						<div class="subsection">
