--- src/data/characterGeneration/data.js
@@ -1,251 +1,0 @@
-// List of available genders
-export const genders = ['Male', 'Female'];
-
-// List of available character classes
-export const classes = [
-	'Barbarian',
-	'Bard',
-	'Cleric',
-	'Druid',
-	'Fighter',
-	'Monk',
-	'Paladin',
-	'Ranger',
-	'Rogue',
-	'Sorcerer',
-	'Warlock',
-	'Wizard'
-];
-
-// List of available alignments
-export const alignments = [
-	'Chaotic Evil',
-	'Lawful Evil',
-	'Neutral Evil',
-	'Neutral',
-	'Neutral Good',
-	'Lawful Good',
-	'Chaotic Good'
-];
-
-// Descriptions of character occupations and their origins
-export const occupations = {
-	Academic: [
-		'Passionate about knowledge and research from a young age',
-		'Inspired by a great teacher or mentor',
-		'Wanted to follow in the footsteps of a family member who was an academic'
-	],
-	Adventurer: [
-		'Craving for excitement and adrenaline rushes',
-		'Love for exploring new and unknown places',
-		'Following the footsteps of a famous adventurer they idolized'
-	],
-	Aristocrat: [
-		'Inherited the family title and fortune',
-		'Wanted to continue the family legacy and maintain their status',
-		'Felt a sense of responsibility to use their privilege to help others'
-	],
-	Artisan: [
-		'Naturally gifted with artistic skills from a young age',
-		'Inspired by a particular work of art or artist',
-		'Found fulfillment in creating beautiful and meaningful things'
-	],
-	'Guild Member': [
-		'Wanted to join a particular guild that aligned with their values and interests',
-		'Needed to learn a trade or skill to support their family',
-		'Was born into a family of guild members and wanted to continue the tradition'
-	],
-	Criminal: [
-		'Grew up in poverty and turned to crime as a means of survival',
-		'Influenced by a bad crowd or family member',
-		'Felt the thrill and power of breaking the law'
-	],
-	Entertainer: [
-		'Naturally talented in singing, dancing, or acting',
-		'Inspired by a particular performer they idolized',
-		'Found joy in making people happy and forgetting their troubles'
-	],
-	Exile: [
-		'Forced to leave their homeland due to political or personal reasons',
-		'Refused to conform to societal norms and was exiled as a result',
-		'Chose to leave voluntarily in search of a better life or adventure'
-	],
-	hermit: [
-		'Traumatized by a past event and wanted to live in seclusion',
-		'Prefers solitude and introspection over socializing',
-		'Wanted to escape the pressures and stress of modern society'
-	],
-	Refugee: [
-		'Fled their home country due to war, persecution, or natural disaster',
-		'Was offered asylum in a new country',
-		'Wanted to start a new life in a place with more opportunities and freedoms'
-	],
-	Explorer: [
-		'Passionate about discovering new and unknown places',
-		'Inspired by a particular explorer they idolized',
-		'Wanted to challenge themselves and push their limits'
-	],
-	Wanderer: [
-		'Wanted to experience new cultures and ways of life',
-		'Felt restless and unsatisfied with their current situation',
-		'Needed to escape a traumatic event or personal problem'
-	],
-	Farmer: [
-		'Comes from a family of farmers and wanted to continue the tradition',
-		'Felt a connection to nature and wanted to work with the land',
-		'Wanted to provide food and sustenance for their community'
-	],
-	Herder: [
-		'Grew up on a farm surrounded by animals and developed a deep love and understanding for them',
-		'Feels a spiritual connection to nature and enjoys being outdoors',
-		'Believes that herding is an important part of preserving traditional ways of life'
-	],
-	Hunter: [
-		'Has always been fascinated by the wild and enjoys the thrill of the hunt',
-		'Believes that hunting is an important part of controlling animal populations',
-		'Is passionate about conservation and wants to ensure that wildlife is protected for future generations'
-	],
-	Trapper: [
-		'Grew up learning trapping skills from family members',
-		'Is fascinated by the challenge of outsmarting wild animals',
-		'Believes that trapping is an important part of wildlife management'
-	],
-	Laborer: [
-		'Comes from a family of blue-collar workers and is proud of their roots',
-		'Believes that hard work is the key to success',
-		'Enjoys the satisfaction of seeing a job well done'
-	],
-	Merchant: [
-		'Has always had a knack for business and enjoys making deals',
-		'Wants to provide for their family and build a legacy',
-		'Is passionate about bringing new products and ideas to their community'
-	],
-	Politician: [
-		'Is deeply committed to social justice and wants to make a difference in the world',
-		'Has a natural charisma and enjoys persuading others to see their point of view',
-		'Believes that government is the best way to effect positive change'
-	],
-	Bureaucrat: [
-		'Is passionate about ensuring that laws are applied fairly and consistently',
-		'Enjoys the challenge of navigating complex systems and regulations',
-		'Believes that bureaucracy is a necessary part of a functioning society'
-	],
-	Priest: [
-		'Feels called to serve a higher power and help others on their spiritual journeys',
-		'Believes that religion can provide comfort and guidance to people in times of need',
-		'Is passionate about promoting love and understanding among people of different faiths'
-	],
-	Sailor: [
-		'Is drawn to the freedom and adventure of life at sea',
-		'Enjoys the challenge of navigating difficult waters and weather conditions',
-		'Believes that trade and commerce are important for building relationships between different cultures and nations'
-	],
-	Soldier: [
-		'Is deeply committed to protecting their country and its people',
-		'Enjoys the camaraderie and discipline of military life',
-		'Believes that serving in the military is a noble and honorable profession'
-	]
-};
-
-// Character backgrounds and related story hooks
-// export const backgrounds = {
-// 	Acolyte: [
-// 		'Seeking to uncover ancient texts and artifacts to uncover the secrets of their deity',
-// 		'Fleeing from their former temple after being framed for a crime they didn\'t commit',
-// 		'On a pilgrimage to spread the teachings of their deity to new lands'
-// 	],
-// 	Charlatan: [
-// 		'Conning their way into high society to seek revenge on a noble who wronged them in the past',
-// 		'Faking their own death to escape the wrath of a powerful crime lord',
-// 		'Trying to swindle a wealthy merchant out of a priceless artifact'
-// 	],
-// 	Criminal: [
-// 		'Seeking to gain enough wealth to retire and live a life of luxury',
-// 		'Attempting to infiltrate a powerful thieves\' guild to take them down from the inside',
-// 		'Framed for a crime they didn\'t commit and seeking to clear their name'
-// 	],
-// 	Entertainer: [
-// 		'Searching for inspiration for their next great performance',
-// 		'On the run from a vengeful ex-lover who is also a powerful crime lord',
-// 		'Trying to save their troupe from being disbanded by a corrupt city official'
-// 	],
-// 	'Folk Hero': [
-// 		'Seeking to overthrow a tyrannical ruler and restore peace to their homeland',
-// 		'Trying to protect their village from a group of bandits who have been terrorizing the area',
-// 		'Rescuing a group of children who have been kidnapped by a nearby tribe'
-// 	],
-// 	'Guild Artisan': [
-// 		'Attempting to create a revolutionary new invention that will change the world',
-// 		'Seeking to uncover a rival guild\'s secrets and gain a competitive advantage',
-// 		'Trying to establish a new guild in an area where there is none'
-// 	],
-// 	Hermit: [
-// 		'Retreating to the wilderness to contemplate the meaning of life and the universe',
-// 		'Fleeing from a powerful enemy who seeks to destroy them and everything they stand for',
-// 		'Trying to uncover a long-lost magical artifact that is rumored to be hidden in the wilderness'
-// 	],
-// 	Noble: [
-// 		'Trying to maintain their family\'s political power and influence in the face of mounting challenges',
-// 		'Seeking to uncover a conspiracy that threatens to undermine the stability of their kingdom',
-// 		'Trying to find a suitable spouse to continue their family\'s legacy'
-// 	],
-// 	Outlander: [
-// 		'Seeking to protect their homeland from an invading army',
-// 		'Trying to uncover the truth about their parents, who disappeared under mysterious circumstances',
-// 		'On a journey to find a rare and valuable resource that is only found in a distant land'
-// 	],
-// 	Sage: [
-// 		'Trying to uncover the true nature of magic and the arcane',
-// 		'Seeking to decode an ancient text that contains forbidden knowledge',
-// 		'Attempting to create a new form of magic that will revolutionize the world'
-// 	],
-// 	Sailor: [
-// 		'Trying to find a new trade route that will make them rich beyond their wildest dreams',
-// 		'Seeking revenge against a rival captain who betrayed them and left them for dead',
-// 		'Rescuing a group of passengers from a sinking ship in the middle of a storm'
-// 	],
-// 	Soldier: [
-// 		'Seeking to restore honor to their name after being dishonorably discharged from the army',
-// 		'Trying to rescue their fellow soldiers who have been captured by an enemy nation',
-// 		'Protecting a small village from an invading army'
-// 	],
-// 	Urchin: [
-// 		'Trying to survive on the streets of a dangerous city',
-// 		'Seeking to avenge the death of their mentor, who was killed by a powerful crime lord',
-// 		'Trying to find a new family after being orphaned at a young age'
-// 	]
-// };
-
-// Different family backgrounds for characters
-export const familyBackgrounds = {
-	'Homeless ': ['Pan handling in the streets '],
-	Entertainers: ['Bards,Minstrels, Poets, Actors '],
-	'Serfs/Servants': ['In service to a feudal lord'],
-	'Free Laborers': ['Self sufficient farmers, fishermen or other subsistence trade'],
-	'Monks or Scholars': ['Parents were immersed in a higher calling'],
-	Military: ['Career soldiers or mercenaries'],
-	Tradesman: ['Practiced a minor trade'],
-	Guild: ['Practiced a popular trade as members of a broad guild'],
-	Merchant: ['From peddlers to gypsies, trade in everything and anything'],
-	'Minor Noble': ['Part of a knight\'s, minor liege lord or other minor household'],
-	'Major Noble': ['Member of a recognized noble family with ties to dukes and barons'],
-	Royalty: ['In the same bloodline as the ruling family']
-};
-
-// Different caretaker origins for characters
-export const caretakersOrigins = {
-	'Original Parents': ['Raised by the ones that gave birth to you'],
-	'Close Family': [
-		'Raised by family members, not your parents (aunt/uncle, grandfather/mother, etc)'
-	],
-	Adopted: ['Never knew original parents. Raised by a couple not related to you.'],
-	Institution: [
-		'Raised at an institution, be it an orphanage, monastery or other.' //Skip to Significant Times in Your Life
-	],
-	Master: [
-		'Sold or given or kidnapped at an early age and raised as property.' //Skip to Significant Times in Your Life
-	],
-	'On Your Own': [
-		'Have had to rely on yourself for as long as you can remember. ' //Skip to Significant Times in Your Life
-	]
-};
