+++ src/data/characterGeneration/misc/alignments.js
@@ -1,0 +1,10 @@
+// List of available alignments
+export const alignments = [
+	'Chaotic Evil',
+	'Lawful Evil',
+	'Neutral Evil',
+	'Neutral',
+	'Neutral Good',
+	'Lawful Good',
+	'Chaotic Good'
+];
