+++ src/data/characterGeneration/misc/classes.js
@@ -1,0 +1,15 @@
+// List of available character classes
+export const classes = [
+	'Barbarian',
+	'Bard',
+	'Cleric',
+	'Druid',
+	'Fighter',
+	'Monk',
+	'Paladin',
+	'Ranger',
+	'Rogue',
+	'Sorcerer',
+	'Warlock',
+	'Wizard'
+];
